Feat: 체커 텍스쳐 파싱시 주기가 짝수인지 확인

+++ includes/texture.h
@@ -34,7 +34,7 @@ typedef struct s_texture_checker
 	t_texture_get_color_at	get_color_at;
 	t_texture				*a;
 	t_texture				*b;
-	double					freq;
+	int						freq;
 }	t_texture_checker;
 
 typedef struct s_texture_image
@@ -50,7 +50,7 @@ void		texture_destroy(t_texture *self);
 
 t_texture	*texture_solid_create(const char *name, t_color color);
 t_texture	*texture_checker_create(const char *name,
-				t_texture *a, t_texture *b, double freq);
+				t_texture *a, t_texture *b, int freq);
 t_texture	*texture_image_create(const char *name,
 				char *imgpath, int rotation, int flip);
